feature: add TypedDict types and add de-/serialization layer
I felt that the type conversion and de-/serialization logic was too
scattered throughout the entity classes, so I added the folder serde in
utils/ and moved moved of the type conversion there. This commit does
not pass pylint, but it passes mypy type checking, and it is already
quite a large commit. There is still a lot that can be done to clean up
the code, but it is considerably better with the introduction of new
type aliases. This commit also reduces redundancy in the tracking/prompt
items by using something akin to a strategy pattern. Again, there is
room for improvement, but it is much better than it was.